+++ tests/config/app.php
@@ -1,11 +1,10 @@
 <?php
 
 return [
-    'storage' => [
-        'cache.default'     => 'file',
-        'cache.stores.file' => [
-            'driver' => 'file',
-            'path'   => 'cache'
+    'cache' => [
+        'driver'  => 'file',
+        'options' => [
+            'path' => 'cache'
         ]
     ]
 ];